refactor: :recycle: Updated all cards, loader and services
Updated all cards, loader, store and services
