+++ components/molecules/cards/index.ts
@@ -1,0 +1,4 @@
+export * from "./filmCard";
+export * from "./personCard";
+export * from "./specieCard";
+export * from "./cardsLoader";
