+++ components/molecules/cards/personCard/index.ts
