+++ components/molecules/cards/specieCard/specieCard.tsx
