+++ components/molecules/index.ts
@@ -2,6 +2,4 @@ export * from "./logo";
 export * from "./header";
 export * from "./navigationItems";
 
-export * from "./filmCard";
-export * from "./personCard";
-export * from "./specieCard";
+export * from "./cards";
