+++ lib/redux/store.ts
@@ -1,12 +1,11 @@
 import { configureStore } from "@reduxjs/toolkit";
 import { setupListeners } from "@reduxjs/toolkit/dist/query";
-import categoryReducer from "./features/categorySlice";
+
 import favouriteReducer from "./features/favouriteSlice";
 import { swaApi } from "../services/swaApi";
 
 export const store = configureStore({
   reducer: {
-    categoryReducer,
     favouriteReducer,
     [swaApi.reducerPath]: swaApi.reducer,
   },
