+++ lib/types/data.ts
@@ -11,10 +11,3 @@ export const DataSchema = z.object({
 });
 
 export type Data = z.infer<typeof DataSchema>;
-
-// TODO: Figure out this shit with an ENUM!!
-export enum CategoryType {
-  FILMS = "films",
-  PEOPLE = "people",
-  SPECIES = "species",
-}
